Add an API around results storage and merge `analytics` with `reporting`
